requirements.txt
tensorflow>=2.6.0
numpy>=1.19.5
matplotlib>=3.3.4
scikit-learn>=0.24.2
pillow>=8.2.0
opencv-python>=4.5.3
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
pillow==8.3.1
tensorflow==2.6.0
numpy==1.19.5
jinja2==3.0.1
aiofiles==0.7.0